chore(tests): Use more recent syntax (02_fib.rs)
