--- tests/should_succeed/cell/02/proof.json
@@ -1,53 +1,0 @@
-{
-  "profile": [
-    { "prover": "z3@4.12.4", "size": 33, "time": 0.461 },
-    { "prover": "cvc5@1.0.5", "size": 46, "time": 0.578 },
-    { "prover": "alt-ergo@2.6.0", "size": 17, "time": 0.512 },
-    { "prover": "cvc4@1.8", "size": 45, "time": 0.39 }
-  ],
-  "proofs": {
-    "M_02__fib": { "vc_fib": { "prover": "cvc5@1.0.5", "time": 0.028 } },
-    "M_02__fib_memo": {
-      "vc_fib_memo": {
-        "tactic": "split_vc",
-        "children": [
-          { "prover": "cvc5@1.0.5", "time": 0.026 },
-          { "prover": "cvc5@1.0.5", "time": 0.039 },
-          { "prover": "cvc5@1.0.5", "time": 0.021 },
-          { "prover": "cvc5@1.0.5", "time": 0.029 },
-          { "prover": "cvc5@1.0.5", "time": 0.037 },
-          { "prover": "cvc5@1.0.5", "time": 0.072 },
-          { "prover": "cvc5@1.0.5", "time": 0.023 },
-          { "prover": "cvc5@1.0.5", "time": 0.02 },
-          { "prover": "cvc5@1.0.5", "time": 0.03 },
-          { "prover": "cvc5@1.0.5", "time": 0.184 },
-          { "prover": "cvc5@1.0.5", "time": 0.037 },
-          { "prover": "cvc5@1.0.5", "time": 0.022 },
-          {
-            "tactic": "split_vc",
-            "children": [
-              {
-                "tactic": "split_vc",
-                "children": [
-                  { "prover": "alt-ergo@2.6.0", "time": 0.028 },
-                  { "prover": "cvc5@1.0.5", "time": 0.034 }
-                ]
-              },
-              { "prover": "cvc5@1.0.5", "time": 0.035 }
-            ]
-          },
-          { "prover": "cvc5@1.0.5", "time": 0.032 },
-          { "prover": "cvc5@1.0.5", "time": 0.05 },
-          { "prover": "cvc5@1.0.5", "time": 0.033 }
-        ]
-      },
-      "vc_get": { "prover": "cvc5@1.0.5", "time": 0.024 },
-      "vc_index": { "prover": "cvc5@1.0.5", "time": 0.017 },
-      "vc_set": { "prover": "cvc5@1.0.5", "time": 0.018 },
-      "vc_v_Some": { "prover": "cvc5@1.0.5", "time": 0.017 }
-    },
-    "M_02__lemma_fib_bound": {
-      "vc_lemma_fib_bound": { "prover": "cvc5@1.0.5", "time": 0.052 }
-    }
-  }
-}
