+++ tests/should_succeed/cell/02_fib/proof.json
@@ -1,0 +1,53 @@
+{
+  "profile": [
+    { "prover": "z3@4.12.4", "size": 33, "time": 0.623 },
+    { "prover": "cvc5@1.0.5", "size": 46, "time": 0.533 },
+    { "prover": "alt-ergo@2.6.0", "size": 17, "time": 0.61 },
+    { "prover": "cvc4@1.8", "size": 45, "time": 0.426 }
+  ],
+  "proofs": {
+    "M_02_fib__fib": { "vc_fib": { "prover": "cvc5@1.0.5", "time": 0.019 } },
+    "M_02_fib__fib_memo": {
+      "vc_fib_memo": {
+        "tactic": "split_vc",
+        "children": [
+          { "prover": "cvc5@1.0.5", "time": 0.04 },
+          { "prover": "cvc5@1.0.5", "time": 0.169 },
+          { "prover": "cvc5@1.0.5", "time": 0.032 },
+          { "prover": "cvc5@1.0.5", "time": 0.058 },
+          { "prover": "cvc5@1.0.5", "time": 0.055 },
+          { "prover": "cvc4@1.8", "time": 0.023 },
+          { "prover": "cvc5@1.0.5", "time": 0.041 },
+          { "prover": "cvc5@1.0.5", "time": 0.055 },
+          { "prover": "cvc5@1.0.5", "time": 0.033 },
+          { "prover": "alt-ergo@2.6.0", "time": 0.167 },
+          { "prover": "cvc4@1.8", "time": 0.032 },
+          { "prover": "cvc5@1.0.5", "time": 0.04 },
+          {
+            "tactic": "split_vc",
+            "children": [
+              {
+                "tactic": "split_vc",
+                "children": [
+                  { "prover": "alt-ergo@2.6.0", "time": 0.159 },
+                  { "prover": "cvc5@1.0.5", "time": 0.168 }
+                ]
+              },
+              { "prover": "cvc5@1.0.5", "time": 0.161 }
+            ]
+          },
+          { "prover": "cvc5@1.0.5", "time": 0.04 },
+          { "prover": "cvc4@1.8", "time": 0.032 },
+          { "prover": "cvc5@1.0.5", "time": 0.04 }
+        ]
+      },
+      "vc_get": { "prover": "cvc5@1.0.5", "time": 0.027 },
+      "vc_index": { "prover": "cvc5@1.0.5", "time": 0.014 },
+      "vc_set": { "prover": "cvc5@1.0.5", "time": 0.02 },
+      "vc_v_Some": { "prover": "cvc5@1.0.5", "time": 0.027 }
+    },
+    "M_02_fib__lemma_fib_bound": {
+      "vc_lemma_fib_bound": { "prover": "cvc5@1.0.5", "time": 0.094 }
+    }
+  }
+}
